Move OneSignal into `onesignal` directory
