+++ src/onesignal/OneSignal.ts
@@ -1,55 +1,55 @@
 import bowser from "bowser";
-import { AuthHashOptionsValidatorHelper } from "./page/helpers/AuthHashOptionsValidatorHelper";
-import { EnvironmentInfoHelper } from "./page/helpers/EnvironmentInfoHelper";
-import AltOriginManager from "./page/managers/AltOriginManager";
-import ConfigManager from "./page/managers/ConfigManager";
-import LegacyManager from "./page/managers/LegacyManager";
-import { AutoPromptOptions } from "./page/managers/PromptsManager";
-import Context from "./page/models/Context";
-import { EnvironmentInfo } from "./page/models/EnvironmentInfo";
-import { NotificationActionButton } from "./page/models/NotificationActionButton";
-import { SecondaryChannelDeviceRecord } from "./shared/models/SecondaryChannelDeviceRecord";
-import { TagsObject } from "./page/models/Tags";
-import ProxyFrame from "./page/modules/frames/ProxyFrame";
-import ProxyFrameHost from "./page/modules/frames/ProxyFrameHost";
-import SubscriptionModal from "./page/modules/frames/SubscriptionModal";
-import SubscriptionModalHost from "./page/modules/frames/SubscriptionModalHost";
-import SubscriptionPopup from "./page/modules/frames/SubscriptionPopup";
-import SubscriptionPopupHost from "./page/modules/frames/SubscriptionPopupHost";
-import TimedLocalStorage from "./page/modules/TimedLocalStorage";
-import Slidedown from "./page/slidedown/Slidedown";
-import { ProcessOneSignalPushCalls } from "./page/utils/ProcessOneSignalPushCalls";
-import { ValidatorUtils } from "./page/utils/ValidatorUtils";
-import OneSignalApi from "./shared/api/OneSignalApi";
-import { InvalidArgumentError, InvalidArgumentReason } from "./shared/errors/InvalidArgumentError";
-import { InvalidStateError, InvalidStateReason } from "./shared/errors/InvalidStateError";
-import { NotSubscribedError, NotSubscribedReason } from "./shared/errors/NotSubscribedError";
-import { SdkInitError, SdkInitErrorKind } from "./shared/errors/SdkInitError";
-import Environment from "./shared/helpers/Environment";
-import EventHelper from "./shared/helpers/EventHelper";
-import HttpHelper from "./shared/helpers/HttpHelper";
-import InitHelper, { RegisterOptions } from "./shared/helpers/InitHelper";
-import MainHelper from "./shared/helpers/MainHelper";
-import OutcomesHelper from "./shared/helpers/OutcomesHelper";
-import SubscriptionHelper from "./shared/helpers/SubscriptionHelper";
-import Emitter, { EventHandler } from "./shared/libraries/Emitter";
-import Log from "./shared/libraries/Log";
-import SdkEnvironment from "./shared/managers/SdkEnvironment";
-import { SessionManager } from "./shared/managers/sessionManager/SessionManager";
-import { AppUserConfig, AppConfig } from "./shared/models/AppConfig";
-import { DeviceRecord } from "./shared/models/DeviceRecord";
-import { NotificationPermission } from "./shared/models/NotificationPermission";
-import { OutcomeAttributionType } from "./shared/models/Outcomes";
-import { DelayedPromptType, AppUserConfigNotifyButton } from "./shared/models/Prompts";
-import { UpdatePlayerOptions } from "./shared/models/UpdatePlayerOptions";
-import { WindowEnvironmentKind } from "./shared/models/WindowEnvironmentKind";
-import Database from "./shared/services/Database";
-import IndexedDb from "./shared/services/IndexedDb";
-import LimitStore from "./shared/services/LimitStore";
-import LocalStorage from "./shared/utils/LocalStorage";
-import OneSignalUtils from "./shared/utils/OneSignalUtils";
-import { awaitOneSignalInitAndSupported, logMethodCall, isValidEmail, getConsoleStyle, executeCallback } from "./shared/utils/utils";
-import OneSignalEvent from "./shared/services/OneSignalEvent";
+import { AuthHashOptionsValidatorHelper } from "../page/helpers/AuthHashOptionsValidatorHelper";
+import { EnvironmentInfoHelper } from "../page/helpers/EnvironmentInfoHelper";
+import AltOriginManager from "../page/managers/AltOriginManager";
+import ConfigManager from "../page/managers/ConfigManager";
+import LegacyManager from "../page/managers/LegacyManager";
+import { AutoPromptOptions } from "../page/managers/PromptsManager";
+import Context from "../page/models/Context";
+import { EnvironmentInfo } from "../page/models/EnvironmentInfo";
+import { NotificationActionButton } from "../page/models/NotificationActionButton";
+import { SecondaryChannelDeviceRecord } from "../shared/models/SecondaryChannelDeviceRecord";
+import { TagsObject } from "../page/models/Tags";
+import ProxyFrame from "../page/modules/frames/ProxyFrame";
+import ProxyFrameHost from "../page/modules/frames/ProxyFrameHost";
+import SubscriptionModal from "../page/modules/frames/SubscriptionModal";
+import SubscriptionModalHost from "../page/modules/frames/SubscriptionModalHost";
+import SubscriptionPopup from "../page/modules/frames/SubscriptionPopup";
+import SubscriptionPopupHost from "../page/modules/frames/SubscriptionPopupHost";
+import TimedLocalStorage from "../page/modules/TimedLocalStorage";
+import Slidedown from "../page/slidedown/Slidedown";
+import { ProcessOneSignalPushCalls } from "../page/utils/ProcessOneSignalPushCalls";
+import { ValidatorUtils } from "../page/utils/ValidatorUtils";
+import OneSignalApi from "../shared/api/OneSignalApi";
+import { InvalidArgumentError, InvalidArgumentReason } from "../shared/errors/InvalidArgumentError";
+import { InvalidStateError, InvalidStateReason } from "../shared/errors/InvalidStateError";
+import { NotSubscribedError, NotSubscribedReason } from "../shared/errors/NotSubscribedError";
+import { SdkInitError, SdkInitErrorKind } from "../shared/errors/SdkInitError";
+import Environment from "../shared/helpers/Environment";
+import EventHelper from "../shared/helpers/EventHelper";
+import HttpHelper from "../shared/helpers/HttpHelper";
+import InitHelper, { RegisterOptions } from "../shared/helpers/InitHelper";
+import MainHelper from "../shared/helpers/MainHelper";
+import OutcomesHelper from "../shared/helpers/OutcomesHelper";
+import SubscriptionHelper from "../shared/helpers/SubscriptionHelper";
+import Emitter, { EventHandler } from "../shared/libraries/Emitter";
+import Log from "../shared/libraries/Log";
+import SdkEnvironment from "../shared/managers/SdkEnvironment";
+import { SessionManager } from "../shared/managers/sessionManager/SessionManager";
+import { AppUserConfig, AppConfig } from "../shared/models/AppConfig";
+import { DeviceRecord } from "../shared/models/DeviceRecord";
+import { NotificationPermission } from "../shared/models/NotificationPermission";
+import { OutcomeAttributionType } from "../shared/models/Outcomes";
+import { DelayedPromptType, AppUserConfigNotifyButton } from "../shared/models/Prompts";
+import { UpdatePlayerOptions } from "../shared/models/UpdatePlayerOptions";
+import { WindowEnvironmentKind } from "../shared/models/WindowEnvironmentKind";
+import Database from "../shared/services/Database";
+import IndexedDb from "../shared/services/IndexedDb";
+import LimitStore from "../shared/services/LimitStore";
+import LocalStorage from "../shared/utils/LocalStorage";
+import OneSignalUtils from "../shared/utils/OneSignalUtils";
+import { awaitOneSignalInitAndSupported, logMethodCall, isValidEmail, getConsoleStyle, executeCallback } from "../shared/utils/utils";
+import OneSignalEvent from "../shared/services/OneSignalEvent";
 
 export default class OneSignal {
   /**
