+++ test/unit/managers/LegacyManager.ts
@@ -3,7 +3,7 @@ import test from "ava";
 import sinon from "sinon";
 import SdkEnvironment from '../../../src/shared/managers/SdkEnvironment';
 import { WindowEnvironmentKind } from '../../../src/shared/models/WindowEnvironmentKind';
-import OneSignal from "../../../src/OneSignal";
+import OneSignal from "../../../src/onesignal/OneSignal";
 
 
 test('OneSignal.environment.getEnv()', async t => {
