+++ test/unit/managers/PermissionManager.ts
@@ -1,7 +1,7 @@
 import '../../support/polyfills/polyfills';
 import test from 'ava';
 import { TestEnvironment, HttpHttpsEnvironment, BrowserUserAgent } from '../../support/sdk/TestEnvironment';
-import OneSignal from '../../../src/OneSignal';
+import OneSignal from '../../../src/onesignal/OneSignal';
 import sinon from 'sinon';
 import Context from '../../../src/page/models/Context';
 import { NotificationPermission } from '../../../src/shared/models/NotificationPermission';
