+++ test/unit/public-sdk-apis/setDefaultNotificationUrl.ts
@@ -3,7 +3,7 @@ import test from "ava";
 import Database from "../../../src/shared/services/Database";
 import Macros from "../../support/tester/Macros";
 import { TestEnvironment } from "../../support/sdk/TestEnvironment";
-import OneSignal from "../../../src/OneSignal";
+import OneSignal from "../../../src/onesignal/OneSignal";
 
 test("url cannot be null",
      Macros.expectInvalidArgumentError,
