+++ test/unit/public-sdk-apis/setDefaultTitle.ts
@@ -2,7 +2,7 @@ import "../../support/polyfills/polyfills";
 import test from "ava";
 import Database from "../../../src/shared/services/Database";
 import { TestEnvironment } from "../../support/sdk/TestEnvironment";
-import OneSignal from "../../../src/OneSignal";
+import OneSignal from "../../../src/onesignal/OneSignal";
 
 
 test("title can be null", async t => {
