+++ test/unit/public-sdk-apis/setEmail.ts
@@ -2,7 +2,7 @@ import "../../support/polyfills/polyfills";
 import test, { ExecutionContext } from "ava";
 import Database from "../../../src/shared/services/Database";
 import { TestEnvironment, BrowserUserAgent } from "../../support/sdk/TestEnvironment";
-import OneSignal from "../../../src/OneSignal";
+import OneSignal from "../../../src/onesignal/OneSignal";
 
 import { InvalidArgumentError } from '../../../src/shared/errors/InvalidArgumentError';
 import nock from 'nock';
