+++ test/unit/public-sdk-apis/showHttpPrompt.ts
@@ -1,5 +1,5 @@
 import test, { ExecutionContext } from "ava";
-import OneSignal from "../../../src/OneSignal";
+import OneSignal from "../../../src/onesignal/OneSignal";
 import { TestEnvironment } from "../../support/sdk/TestEnvironment";
 import Context from "../../../src/page/models/Context";
 import sinon, { SinonSandbox } from 'sinon';
